backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
graphene
psycopg2-binary
python-dotenv
PyJWT